refactor(promotion)!: remove priority queue
